requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blessings==1.7
cachetools==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fire==0.4.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==2.6.2
gpustat==0.6.0
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jill==0.11.0
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.0
jupyter-core==4.9.2
kiwisolver==1.4.2
markdown==3.3.6
matplotlib-inline==0.1.3
matplotlib==3.5.2
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.3
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.1.0
pip==22.0.4
portalocker==2.4.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydeprecate==0.3.1
pygments==2.12.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-gnupg==0.4.9
pytorch-lightning==1.9.3
pyyaml==6.0
pyzmq==22.3.0
requests-futures==1.0.0
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.8
sacrebleu==1.4.14
scikit-learn==1.0.2
scipy==1.8.0
semantic-version==2.10.0
sentencepiece==0.1.94
sentry-sdk==1.5.8
setproctitle==1.2.2
setuptools==59.5.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
termcolor==1.1.0
threadpoolctl==3.1.0
torchmetrics==1.0.3 
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
werkzeug==2.1.1
wget==3.2
wheel==0.37.1
yarl==1.7.2
zipp==3.8.0